+++ CHANGELOG.md
@@ -36,6 +36,7 @@ UIIN-3437.
 * Display default value in Language select when language is incorrect. Fixes UIIN-3451.
 * Correcting the behavior of drop-down menu and checkbox in Settings > Inventory > Number generator options. Refs UIIN-3471.
 * Move permissions for creating export jobs to "Inventory: View instances, holdings, and items" permission. Fixes UIIN-3474.
+* Fix sorting for Items on Instance details page. Fixes UIIN-3478.
 * ECS: Set Held by facet default to current tenant context in Inventory Search. Refs UIIN-2820.
 
 ## [13.0.9](https://github.com/folio-org/ui-inventory/tree/v13.0.9) (2025-08-26)