+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-inventory
 
+## [13.0.10] (IN PROGRESS)
+
+* Fix default staff suppress facet value issues after a merge error. Refs UIIN-3465.
+
 ## [13.0.9](https://github.com/folio-org/ui-inventory/tree/v13.0.9) (2025-08-26)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.8...v13.0.9)